k8s: Extend Node subscriber to accept swg
The swg (stoppable wait group) is used by the service_cache.go when
syncing k8s caches upon the agent startup. Until now, service_cache was
consuming only Service and Endpoint* objects. However, for the upcoming
service topology aware hints feature we need to add (self) Node object
as well to the list. This is because the feature needs to get the
"topology.kubernetes.io/zone" of the self Node.

Signed-off-by: Martynas Pumputis <m@lambda.lt>